requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
openai==0.28.1
cloudinary==1.36.0
Pillow==10.0.1
python-multipart==0.0.6
